requirements.txt
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
defusedxml==0.7.1
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.9.0
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
-e git+https://github.com/IDEA-Research/GroundingDINO.git@27024e42dac6121738d396f1f2b0dc1c9093e9d4#egg=groundingdino
h11==0.14.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==7.0.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
-e git+https://github.com/sachit-menon/GLIP.git@d40518186100d86eff8718ecd09755c7a0418186#egg=maskrcnn_benchmark&subdirectory=../../GLIP
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pycocotools==2.0.7
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.1
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
supervision==0.18.0
sympy==1.12
timm==0.9.12
tokenizers==0.15.2
tomli==2.0.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.2
transformers==4.37.2
triton==2.1.0
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.27.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
